fonksiyon varlığı denetimi eklendi

+++ public/assets/admin/js/custom.js
@@ -6,8 +6,9 @@ $(function () {
 	/**
 	 * dataTable sınıflı  tablolar DataTeble ile düzenleniyor
 	 */
-	$('.dataTable').dataTable();
-
+	if ( $.isFunction($.fn.dataTable) ) {
+		$('.dataTable').dataTable();
+	}
 	/**
 	 * ajaxForm sınıflı  formların işlemini ajax ile yapıyor
 	 */
@@ -119,31 +120,34 @@ $(function () {
 	/**
 	 *  data-mask özelliği olan nesneye input mask uygular
 	 */
-	$("[data-mask]").inputmask();
-
+	if ( $.isFunction($.fn.inputmask) ) {
+		$("[data-mask]").inputmask();
+	}
 	/**
 	 * TinyMCE tinymce sınıflı  textarea üzerine yerleşecek
 	 */
-	tinymce.init({
-		selector                 : "textarea.tinymce",
-		menubar                  : false,
-		language                 : 'tr_TR',
-		theme                    : "modern",
-		plugins                  : [
-			"advlist autolink link image lists charmap print preview hr anchor pagebreak",
-			"searchreplace wordcount visualblocks visualchars insertdatetime media nonbreaking",
-			"table contextmenu directionality emoticons paste textcolor responsivefilemanager",
-			"code spellchecker"
-		],
-		toolbar1                 : "|  preview code undo redo | bold italic underline | alignleft aligncenter alignright alignjustify | bullist numlist outdent indent | styleselect | responsivefilemanager | image media",
-		toolbar2                 : "| link unlink anchor  | forecolor backcolor  | hr link image charmap paste copy cut pagebreak searchreplace | insertdatetime table ltr rtl",
-		image_advtab             : true,
-		external_filemanager_path: "/assets/admin/js/plugins/ResponsiveFilemanager/filemanager/",
-		filemanager_title        : "Responsive Filemanager",
-		external_plugins         : {"filemanager": "/assets/admin/js/plugins/ResponsiveFilemanager/filemanager/plugin.min.js"}
+	if ( $.isFunction($.fn.tinymce) ) {
+		tinymce.init({
+			selector                 : "textarea.tinymce",
+			menubar                  : false,
+			language                 : 'tr_TR',
+			theme                    : "modern",
+			plugins                  : [
+				"advlist autolink link image lists charmap print preview hr anchor pagebreak",
+				"searchreplace wordcount visualblocks visualchars insertdatetime media nonbreaking",
+				"table contextmenu directionality emoticons paste textcolor responsivefilemanager",
+				"code spellchecker"
+			],
+			toolbar1                 : "|  preview code undo redo | bold italic underline | alignleft aligncenter alignright alignjustify | bullist numlist outdent indent | styleselect | responsivefilemanager | image media",
+			toolbar2                 : "| link unlink anchor  | forecolor backcolor  | hr link image charmap paste copy cut pagebreak searchreplace | insertdatetime table ltr rtl",
+			image_advtab             : true,
+			external_filemanager_path: "/assets/admin/js/plugins/ResponsiveFilemanager/filemanager/",
+			filemanager_title        : "Responsive Filemanager",
+			external_plugins         : {"filemanager": "/assets/admin/js/plugins/ResponsiveFilemanager/filemanager/plugin.min.js"}
 
 
-	});
+		});
+	}
 	/**
 	 * şehir seçildiğinde otomatik olark ilçe listesini getirmek için
 	 */
